restruct arrow
