+++ raphtory/src/io/mod.rs
@@ -1,3 +1,6 @@
 pub mod csv_loader;
 pub mod json_loader;
 pub mod neo4j_loader;
+#[cfg(feature = "arrow")]
+pub(crate) mod arrow;
+pub mod parquet_loaders;
